move code into Model.c, target gnu89
