+++ src/Model.h
@@ -4,10 +4,15 @@
 #include "PackedCol.h"
 #include "Constants.h"
 #include "Graphics.h"
+#include "Physics.h"
 /* Contains various structs and methods for an entity model.
    Also contains a list of models and default textures for those models.
    Copyright 2014-2019 ClassiCube | Licensed under BSD-3
 */
+struct Entity;
+struct AABB;
+struct IGameComponent;
+struct VertexTextured;
 extern struct IGameComponent Models_Component;
 
 #define MAX_CUSTOM_MODELS 64
@@ -16,7 +21,6 @@ extern struct IGameComponent Models_Component;
 #define MODEL_QUAD_VERTICES 4
 #define MODEL_BOX_VERTICES (FACE_COUNT * MODEL_QUAD_VERTICES)
 enum RotateOrder { ROTATE_ORDER_ZYX, ROTATE_ORDER_XZY, ROTATE_ORDER_YZX, ROTATE_ORDER_XYZ };
-struct VertexTextured defaultVertices[MODEL_BOX_VERTICES * MAX_CUSTOM_MODEL_PARTS];
 
 /* Describes a vertex within a model. */
 struct ModelVertex { float X, Y, Z; cc_uint16 U, V; };
@@ -146,7 +150,7 @@ CC_API struct Model* Model_Get(const String* name);
 /* Adds a model to the list of models. (e.g. "skeleton") */
 /* Models can be applied to entities to change their appearance. Use Entity_SetModel for that. */
 CC_API void Model_Register(struct Model* model);
-/* Unregister a model from the list of models. */
+/* Unregister a model from the list of models, and set all entities using this model to the default humanoid model. */
 void Model_Unregister(struct Model* model);
 /* Adds a texture to the list of automatically managed model textures. */
 /* These textures are automatically loaded from texture packs. (e.g. "skeleton.png") */
@@ -206,4 +210,62 @@ CC_API void BoxDesc_ZQuad(struct Model* m, int texX, int texY, int texWidth, int
 CC_API void BoxDesc_XQuad2(struct Model* m, float z1, float z2, float y1, float y2, float x, int u1, int v1, int u2, int v2);
 CC_API void BoxDesc_YQuad2(struct Model* m, float x1, float x2, float z1, float z2, float y, int u1, int v1, int u2, int v2);
 CC_API void BoxDesc_ZQuad2(struct Model* m, float x1, float x2, float y1, float y2, float z, int u1, int v1, int u2, int v2);
+
+/* CustomModel types */
+
+enum CustomModelAnim {
+	CustomModelAnim_None = 0,
+	CustomModelAnim_Head = 1,
+	CustomModelAnim_LeftLeg = 2,
+	CustomModelAnim_RightLeg = 3,
+	CustomModelAnim_LeftArm = 4,
+	CustomModelAnim_RightArm = 5,
+};
+
+struct CustomModelPart {
+	struct ModelPart model_part;
+
+	struct BoxDesc boxDesc;
+	float rotationX;
+	float rotationY;
+	float rotationZ;
+	enum CustomModelAnim anim;
+	cc_bool fullbright;
+};
+
+struct CustomModel {
+	struct Model model;
+	struct ModelTex defaultTex;
+
+	char name[STRING_SIZE + 1];
+	struct ModelVertex* vertices;
+	float nameY;
+	float eyeY;
+	Vec3 collisionBounds;
+	struct AABB pickingBoundsAABB;
+
+	cc_bool bobbing;
+	cc_bool pushes;
+	/* if true, falls back to using your account's skin */
+	cc_bool usesHumanSkin;
+	/* use crazy arms */
+	cc_bool calcHumanAnims;
+
+	cc_bool hideFirstPersonArm;
+
+	cc_uint16 uScale;
+	cc_uint16 vScale;
+
+	cc_uint8 numParts;
+	struct CustomModelPart parts[MAX_CUSTOM_MODEL_PARTS];
+
+	cc_bool valid;
+};
+
+static struct CustomModel custom_models[MAX_CUSTOM_MODELS];
+
+void CustomModel_Init(struct CustomModel* customModel);
+void CustomModel_Free(struct CustomModel* customModel);
+void CustomModel_FreeAll(void);
+
 #endif
